Split Container into a generic and a non-generic part
This simplification of the code has surprisingly powerful performance
and convenience implications. IContainer is no longer needed, and
common container methods / properties can now be accessed without
virtual dispatch. Further, due to the simplified casting, it is now
possible to flatten the DrawNode hierarchy of generic containers,
neat! Lastly, due to the avoidance of IContainer some members
can now have non-public visibility, such as InternalChildren et al.
This allows making containers which manage their own children
and disallow adding / removing from outside.
