+++ osu.Framework/Graphics/Containers/IContainer.cs
@@ -7,36 +7,17 @@ using OpenTK;
 
 namespace osu.Framework.Graphics.Containers
 {
-    public interface IContainer : IDrawable
-    {
-        Vector2 ChildSize { get; }
-        Vector2 ChildOffset { get; }
-        Vector2 RelativeToAbsoluteFactor { get; }
-        Vector2 RelativeChildOffset { get; }
-
-        EdgeEffectParameters EdgeEffect { get; set; }
-        float CornerRadius { get; }
-
-        void InvalidateFromChild(Invalidation invalidation);
-
-        void Clear(bool dispose = true);
-
-        Axes AutoSizeAxes { get; set; }
-    }
-
-    public interface IContainerEnumerable<out T> : IContainer
+    public interface IContainerEnumerable<out T>
         where T : IDrawable
     {
-        IReadOnlyList<Drawable> InternalChildren { get; }
         IReadOnlyList<T> Children { get; }
 
         int RemoveAll(Predicate<T> match);
     }
 
-    public interface IContainerCollection<in T> : IContainer
+    public interface IContainerCollection<in T>
         where T : IDrawable
     {
-        IReadOnlyList<Drawable> InternalChildren { set; }
         IReadOnlyList<T> Children { set; }
 
         void Add(T drawable);
