+++ osu.Framework/Graphics/IDrawable.cs
@@ -24,7 +24,7 @@ namespace osu.Framework.Graphics
         /// <summary>
         /// The parent of this drawable in the scene graph.
         /// </summary>
-        IContainer Parent { get; }
+        ContainerBase Parent { get; }
 
         /// <summary>
         /// Whether this drawable is present for any sort of user-interaction.
