+++ osu.Framework/Graphics/Transforms/TransformEdgeEffectAlpha.cs
@@ -10,7 +10,7 @@ namespace osu.Framework.Graphics.Transforms
         public override void Apply(Drawable d)
         {
             base.Apply(d);
-            IContainer c = (IContainer)d;
+            ContainerBase c = (ContainerBase)d;
 
             EdgeEffectParameters e = c.EdgeEffect;
             e.Colour.Linear.A = CurrentValue;
