+++ osu.Framework/Input/InputManager.cs
@@ -694,7 +694,7 @@ namespace osu.Framework.Input
             if (stillValid)
             {
                 //ensure we are visible
-                IContainer d = FocusedDrawable.Parent;
+                ContainerBase d = FocusedDrawable.Parent;
                 while (d != null)
                 {
                     if (!d.IsPresent)
